[#12] Add new field for max character count of abstract
Many of the abstracts submitted have been quite short. To clarify
if this is because of the conference's constraint's or the
speaker's choice, a new field is added: max character count.

This will allow reviewers refrain from suggesting more details
should be added to an abstract if it's the conference which is
requesting the short version of the abstract.

Adjust:

- Form collecting the name, and gist of the person who needs help
- Mail which is sent to volunteer reviewers
- Style to shorten the field for max characters

Resolves issue [#12].