Normalise line endings for text files
CSS file was saved with a different line ending, which caused
problems when I tried to add a new style for the max character
count.

This update normalises the line endings and adds a .gitattributes
file so that future contributors don't need to worry about what
line endings their text editor was using.